apps/leslighet/LixService/requirements.txt
# Web Framework
fastapi== 0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
pydantic==2.11.4
aiofiles==24.1.0

# Text Analysis
textstat==0.7.5

# Data Messaging
sse-starlette==2.3.3
websockets==15.0.1
aio_pika==9.5.5 # RabbitMQ client

# Caching
redis==5.2.1
tenacity==9.1.2

# Monitoring and Metrics
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
psutil==7.0.0

# Utils
python-dotenv==1.1.0
structlog==25.3.0
aiohttp==3.11.18
requests==2.32.3
httpx==0.28.1


# Code Quality (optional for running)
black==25.1.0
isort==6.0.1
flake8==7.2.0
mypy==1.15.0